fix: MSVC /W4 real-bug cleanup + per-class suppressions (Phase 5 T9)
Census from CI run 29208716185 (first /W4 exercise; per-class unique
site counts, both targets deduped). Real fixes in source:

- LNK2001: act_format_tests.cpp declared specialize_name without the
  top-level const of its consts.cpp definition -- MSVC mangles variable
  types into decorated names (GNU/Itanium does not), so the mismatch
  only breaks the Windows link.
- C4700/C4701/C4703 (23 sites): uninitialized/potentially-uninitialized
  locals. Most are flow-analysis false positives (init added, behavior
  unchanged); genuinely reachable indeterminate reads (UB, now
  deterministic, disclosed per constraint): do_gen_com imm_side (mortal
  speaker of an unmatched race + immortal listener), do_butcher no_tool
  (Beorning natural-claws branch never set it), Crash_delete_crashfile
  rent (truncated crash file), run_script tmpint (TELEPORT_CHAR_XL
  reading another case's variable -- stale-reuse bug preserved,
  first-read now 0), get_real_OB tactics (non-enum GET_TACTICS),
  append_room i (read before first fscanf), unaccent 255->'\xff'.
- C4477 (3): fprintf %d/%ld with SocketType (unsigned 64-bit SOCKET on
  Windows) and time_t (64-bit there) -- real varargs mismatches; %lld +
  cast prints identical digits everywhere (boot golden verified).
- C4550 (2): `if (*fptr)` -> `if (fptr)` (identical semantics).
- C4804 (2): MSVC branch added to the existing bug-preservation pragma
  blocks around the preserved always-true chained comparisons.
- C4805 (1): explicit int casts in the bool<<1|bool sort-flag pack.
- C4312 (2, tests): 0xdeadbeef garbage sentinel routed via uintptr_t.

Suppressed per-class with rationale + ledger in src/CMakeLists.txt:
/wd4244 /wd4267 (pervasive benign legacy narrowing, 993+174 sites --
sanctioned this phase, future effort may revisit), /wd4456 /wd4458
/wd4459 (shadowing; GNU bar has no -Wshadow either), /wd4702
(unreachable legacy switch arms), /wd4127 (constant conditionals,
mostly test macros), /wd4310 (telnet byte constants; MSVC's check
ignores /J). C4996 POSIX-name deprecations handled by
_CRT_NONSTDC_NO_WARNINGS alongside the existing secure-CRT knob.

Local gate: macOS arm64 build (-Werror) clean, ctest 1015/1015,
boot-golden verify byte-identical.

+++ src/db.cpp
@@ -2937,14 +2937,14 @@ bool write_player_text(struct char_data* ch, int load_room, const char* scratch_
     fprintf(pf, "bodytype    %d\n", chd.bodytype);
     fprintf(pf, "level       %d\n", chd.level);
     fprintf(pf, "language    %d\n", chd.language);
-    fprintf(pf, "birth       %ld\n", chd.birth);
+    fprintf(pf, "birth       %lld\n", static_cast<long long>(chd.birth));
     fprintf(pf, "played      %d\n", chd.played);
     fprintf(pf, "weight      %d\n", chd.weight);
     fprintf(pf, "height      %d\n", chd.height);
     fprintf(pf, "title       %s\n", chd.title);
     fprintf(pf, "hometown    %d\n", chd.hometown);
     fprintf(pf, "description \n%s~\n", chd.description);
-    fprintf(pf, "last_logon  %ld\n", chd.last_logon);
+    fprintf(pf, "last_logon  %lld\n", static_cast<long long>(chd.last_logon));
     memcpy(pwdcrypt, chd.pwd, MAX_PWD_LENGTH);
     encrypt_line((unsigned char*)pwdcrypt, MAX_PWD_LENGTH);
     pwdcrypt[MAX_PWD_LENGTH] = '\0'; // terminate explicitly -- see the pwdcrypt declaration comment
